Split dashboard into modular components
